pyproject.toml
[tool.poetry]
name = "nonebot-plugin-moellmchats"
version = "0.14.6"
description = "A Nonebot plugin, Multi-Model Chat System | Hybrid Scheduling·Search Augmented·Persona Customization"
license = "GPL"
authors = ["Elflare"]
readme = "README.md"
homepage = "https://github.com/Elflare/nonebot-plugin-moellmchats"
repository = "https://github.com/Elflare/nonebot-plugin-moellmchats"
documentation = "https://github.com/Elflare/nonebot-plugin-moellmchats#readme"
keywords = ["nonebot", "nonebot2", "llm", "ai", "chat"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.1"
nonebot2 = "^2.2.0"
nonebot-adapter-onebot = "^2.0.0"
ujson = "^5.7.0"
nonebot-plugin-localstore = "^0.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
nonemoji = "^0.1.2"
pre-commit = "^4.0.0"

[tool.ruff]
line-length = 130
target-version = "py39"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "TID",   # flake8-tidy-imports
    "TC",    # flake8-type-checking
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [{ root = "./" }]

typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
